requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.14
langgraph
openai
numpy==1.26.4
pandas==2.2.2
sqlalchemy==2.0.28
pymongo==4.6.3
motor==3.4.0
neo4j==5.16.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
httpx==0.27.0